+++ R/pl1_pipeline_main.R
@@ -5,7 +5,8 @@ pl_main <- function(mdat, mode = "rocprc", calc_avg = TRUE, cb_alpha = 0.05,
                     raw_curves = FALSE, x_bins = 1000, interpolate = TRUE,
                     na_worst = TRUE, ties_method = "equiv", beta = 1,
                     on_single_class = "error", metrics = NULL,
-                    cost_fp = 1, cost_fn = 1, validate = TRUE) {
+                    cost_fp = 1, cost_fn = 1, basic_ties = "split",
+                    validate = TRUE) {
   # === Validation ===
   new_mode <- .pmatch_mode(mode)
   on_single_class <- .pmatch_on_single_class(on_single_class)
@@ -13,7 +14,7 @@ pl_main <- function(mdat, mode = "rocprc", calc_avg = TRUE, cb_alpha = 0.05,
     .validate_pl_main_args(
       mdat, new_mode, calc_avg, cb_alpha, raw_curves,
       x_bins, interpolate, beta, on_single_class, metrics,
-      cost_fp, cost_fn
+      cost_fp, cost_fn, basic_ties
     )
   }
 
@@ -32,7 +33,8 @@ pl_main <- function(mdat, mode = "rocprc", calc_avg = TRUE, cb_alpha = 0.05,
     .pl_main_basic(mdat, model_type, dataset_type, class_name_pf,
       calc_avg = calc_avg, cb_alpha = cb_alpha,
       raw_curves = raw_curves, beta = beta, metrics = metrics,
-      cost_fp = cost_fp, cost_fn = cost_fn
+      cost_fp = cost_fp, cost_fn = cost_fn, x_bins = x_bins,
+      basic_ties = basic_ties
     )
   } else if (new_mode == "aucroc") {
     .pl_main_aucroc(mdat, model_type, dataset_type, class_name_pf,
@@ -102,7 +104,7 @@ pl_main <- function(mdat, mode = "rocprc", calc_avg = TRUE, cb_alpha = 0.05,
                                    x_bins, interpolate, beta = 1,
                                    on_single_class = "error",
                                    metrics = NULL, cost_fp = 1,
-                                   cost_fn = 1) {
+                                   cost_fn = 1, basic_ties = "split") {
   # Validate mdat
   .validate(mdat)
   if (mode != "aucroc" && !is.null(mdat) && length(mdat) > 0 &&
@@ -141,6 +143,9 @@ pl_main <- function(mdat, mode = "rocprc", calc_avg = TRUE, cb_alpha = 0.05,
 
   # Check the misclassification costs
   .validate_costs(cost_fp, cost_fn)
+
+  # Check basic_ties
+  .validate_basic_ties(basic_ties)
 }
 
 #